add watch non-namespaced ConfigMaps works!
Implemented via Watch handler allowing a user-defined mapping function.
It will be run for every watch even on configmaps per-se and then allows
us to query for a specific property and based on it decide if our
reconciler shall be triggered.
For now hardcoded the reconciler. Next we need to manage this specific
label or annotation on the configmap and then in the second step
implement parsing form it the configmapsync objects name and namespace!